Implement splitting long texts
