+++ pali_bot/__main__.py
@@ -8,6 +8,7 @@ from pali_bot.sutta_provider import SuttaProvider
 
 def main() -> None:
     logging.basicConfig(level=config.LOG_LEVEL)
+    DIR = os.path.abspath(os.path.dirname(__file__))  # TODO
     sutta_provider = SuttaProvider()  # TODO config
     bot = Bot(sutta_provider=sutta_provider)
     bot.run()
